# src/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/packages": {
            "post": {
                "description": "Create a new package with the provided information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packages"
                ],
                "summary": "Create a new package",
                "parameters": [
                    {
                        "description": "Create Package",
                        "name": "package",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/deploy-buddy_server_internal_model.CreatePackageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/packages/metadata": {
            "post": {
                "description": "List metadata with the provided connection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packages"
                ],
                "summary": "List metadata",
                "parameters": [
                    {
                        "description": "List Metadata",
                        "name": "metadata",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ListMetadatasRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/deploy-buddy_server_internal_repository.MetadataResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/slack/interactive": {
            "post": {
                "description": "Handle interactive messages from Slack.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slack"
                ],
                "summary": "Handle interactive messages from Slack",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "deploy-buddy_server_internal_model.CreatePackageRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "repo_name": {
                    "type": "string"
                },
                "string_connection": {
                    "type": "string"
                }
            }
        },
...
        "deploy-buddy_server_internal_repository.MetadataResult": {
            "type": "object",
            "properties": {
                "createdByName": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.ListMetadatasRequest": {
            "type": "object",
            "properties": {
                "string_connection": {
                    "type": "string"
                }
            }
        }
    }
}